Chore: rename modules for clarity
